refactor: 하단탭 컴포넌트 상수 및 타입 분리
