+++ components/common/BottomTab/BottomTabItem.tsx
@@ -3,24 +3,7 @@
 import Image, { StaticImageData } from 'next/image'
 import { usePathname, useRouter } from 'next/navigation'
 import { home, exp, quest, noticeBoard, user, active_home, active_exp, active_quest, active_noticeBoard, active_user } from '@/assets/images'
-
-type ContentKey = '홈' | '경험치' | '퀘스트' | '게시판' | '사용자';
-
-const pathToContent: Record<string, ContentKey> = {
-  'home': '홈',
-  'exp': '경험치',
-  'quest': '퀘스트',
-  'board': '게시판',
-  'mypage': '사용자'
-};
-
-const contentToPath: Record<ContentKey, string> = {
-  '홈': '/home',
-  '경험치': '/exp',
-  '퀘스트': '/quest',
-  '게시판': '/board',
-  '사용자': '/mypage'
-};
+import { contentToPath, pathToContent } from '@/constant';
 
 const imageMap: Record<ContentKey, { default: StaticImageData, active: StaticImageData }> = {
   '홈': { default: home, active: active_home },
