+++ constant/index.ts
@@ -1,0 +1,2 @@
+export { pathTitle } from './header'
+export { pathToContent, contentToPath } from './bottom-tab-item'
