+++ constant/type.d.ts
@@ -1,0 +1,5 @@
+type pathTitle = {
+  [key: string]: string
+}
+
+type ContentKey = '홈' | '경험치' | '퀘스트' | '게시판' | '사용자';
